Move logic to respective and use index.js to import and export components
